requirements.txt
numpy~=1.19.5
requests~=2.25.1
tensorflow~=2.5.0
Pillow~=8.2.0
Flask~=2.0.1
gunicorn==20.1.0
firebase-admin